#17: namespaceをリファクタリングした
ついでに下記も気がついた分だけやっておいた
- 実装の過程で使用されなくなったuse文の除去
- 使用されていないファイルの除去
- LoginController::inputAction のFeatureテスト(画面生成がコケないことを確認するだけ)が未作成だったので作成
