--- BizLogics/Aggregate/User/Exception/PasswordTypeCompositionInvalidException.php
@@ -1,11 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Bizlogics\Aggregate\User\Exception;
-
-use RuntimeException;
-
-final class PasswordTypeCompositionInvalidException extends RuntimeException
-{
-
-}
