+++ BizLogics/Uam/Aggregate/User/Exception/PasswordSizeTooShortException.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Bizlogics\Uam\Aggregate\User\Exception;
+
+use RuntimeException;
+
+final class PasswordSizeTooShortException extends RuntimeException
+{
+
+}
