+++ BizLogics/Uam/UseCase/UserOperation/CreateAccount/Exception/ApplyingException.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Bizlogics\Uam\UseCase\UserOperation\CreateAccount\Exception;
+
+use RuntimeException;
+
+final class ApplyingException extends RuntimeException
+{
+
+}
