+++ BizLogics/Uam/UseCase/UserOperation/CreateAccount/Exception/EmailAlreadyUsedException.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Bizlogics\Uam\UseCase\UserOperation\CreateAccount\Exception;
+
+use RuntimeException;
+
+final class EmailAlreadyUsedException extends RuntimeException
+{
+
+}
