--- BizLogics/UseCase/UserOperation/CreateAccount/Exception/ApplyingException.php
@@ -1,11 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace Bizlogics\UseCase\UserOperation\CreateAccount\Exception;
-
-use RuntimeException;
-
-final class ApplyingException extends RuntimeException
-{
-
-}
