--- app/Infrastructure/AggregateRepository/Exception/RegistrationProcessFailedException.php
@@ -1,11 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\Infrastructure\AggregateRepository\Exception;
-
-use RuntimeException;
-
-final class RegistrationProcessFailedException extends RuntimeException
-{
-
-}
