--- app/Infrastructure/AggregateRepository/User/Exception/NotExistException.php
@@ -1,9 +1,0 @@
-<?php
-declare(strict_types=1);
-
-namespace App\Infrastructure\AggregateRepository\User\Exception;
-
-final class NotExistException extends \RuntimeException
-{
-
-}
