+++ app/Infrastructure/Uam/AggregateRepository/Exception/RegistrationProcessFailedException.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Infrastructure\Uam\AggregateRepository\Exception;
+
+use RuntimeException;
+
+final class RegistrationProcessFailedException extends RuntimeException
+{
+
+}
