+++ app/Infrastructure/Uam/AggregateRepository/User/Exception/NotExistException.php
@@ -1,0 +1,9 @@
+<?php
+declare(strict_types=1);
+
+namespace App\Infrastructure\Uam\AggregateRepository\User\Exception;
+
+final class NotExistException extends \RuntimeException
+{
+
+}
